requirements.txt
numpy==1.18.2
opencv-contrib-python-headless==4.2.0.34
opencv-python-headless==4.2.0.32
Pillow==7.0.0
PyQt5==5.14.1
PyQt5-sip==12.7.1
python-dateutil==2.8.1
scipy==1.4.1
six==1.14.0
